many things

+++ SilkySouls2/Interfaces/IEventService.cs
@@ -4,8 +4,8 @@ namespace SilkySouls2.Interfaces
 {
     public interface IEventService
     {
-        void SetEvent(long eventId, bool setVal);
-        void SetMultipleEventOn(long[] gameIds);
-        bool GetEvent(long gameId);
+        void SetEvent(int eventId, bool setVal);
+        void SetMultipleEventOn(int[] gameIds);
+        bool GetEvent(int gameId);
     }
 }
